pyproject.toml
[tool.poetry]
name = "sky-bank"
version = "0.1.0"
description = ""
authors = ["elena_pantsurkina <pantsurkina.e@bk.ru>"]
readme = "README.md"
virtualenvs.in-project = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.1"
black = "^24.4.2"
isort = "^5.13.2"
[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

black = "^24.4.2"
[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

isort = "^5.13.2"
[tool.isort]
# максимальная длина строки
line_length = 119



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"